Refactoring code and empty component

+++ src/server/api/routers/conductores.ts
@@ -1,4 +1,9 @@
-import { createTRPCRouter, publicProcedure } from "@/server/api/trpc";
+import { TRPCError } from "@trpc/server";
+import {
+  createTRPCRouter,
+  protectedProcedure,
+  publicProcedure,
+} from "@/server/api/trpc";
 import { conductorSchema } from "@/schemas";
 import { z } from "zod";
 
@@ -26,9 +31,16 @@ export const conductoresRouter = createTRPCRouter({
       }
     }),
 
-  deleteConductor: publicProcedure
+  deleteConductor: protectedProcedure
     .input(z.object({ id: z.string() }))
     .mutation(async ({ input, ctx }) => {
+      const isNotAdmin = ctx.session.user.rol !== "ADMIN";
+      if (isNotAdmin) {
+        throw new TRPCError({
+          code: "UNAUTHORIZED",
+          message: "No tienes permisos para realizar esta acción",
+        });
+      }
       try {
         await ctx.prisma.conductor.delete({
           where: { id: input.id },
